+++ src/atoms/cc-toggle.js
@@ -47,6 +47,7 @@ import { dispatchCustomEvent } from '../lib/events.js';
  * @event {CustomEvent<String>} cc-toggle:input - Fires the selected `value` whenever the selected `value` changes (single mode only).
  * @event {CustomEvent<String[]>} cc-toggle:input-multiple - Fires the selected `multipleValues` whenever the selected `multipleValues` changes (single mode only).
  *
+ * @cssprop {TextTransform} --cc-text-transform - Apply a text transformation on labels (defaults: `uppercase`).
  * @cssprop {Color} --cc-toggle-color - The main color of the toggle (defaults: `#334252`). It must be defined directly on the element.
  * @cssprop {Filter} --cc-toggle-img-filter - A CSS filter to apply on images of all choices (defaults: `none`). It must be defined directly on the element.
  * @cssprop {Filter} --cc-toggle-img-filter-selected - A CSS filter to apply on images of selected choices (defaults: `none`). It must be defined directly on the element.
@@ -213,7 +214,7 @@ export class CcToggle extends LitElement {
           grid-gap: 0.5rem;
           padding: 0 0.5rem;
           position: relative;
-          text-transform: uppercase;
+          text-transform: var(--cc-text-transform, uppercase);
           -moz-user-select: none;
           -webkit-user-select: none;
           -ms-user-select: none;